requirements.txt
altgraph==0.17.4
anyio==4.7.0
certifi==2024.8.30
charset-normalizer==3.4.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
packaging==24.2
pefile==2023.2.7
Pillow==11.0.0
psutil==5.9.8
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.10
pypiwin32==223
pystray==0.19.5
pywin32==308
pywin32-ctypes==0.2.3
requests==2.31.0
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.2.3
win10toast==0.9
nvidia-ml-py
sv_ttk